refactor: rename the globe social media container to live updates
The container has held publications as well as Mastodon posts since the
carousel started mixing them, so "social media" named only half of what it
renders. Renames the directory, the two feed components and the call-site
aliases, and switches the moved files from `../` to `@/` imports.

Also memoizes the feed: `toFeedItems` merges and sorts on every render, and
the mobile feed re-renders on every slide change. Memoizing it there only
works if the source arrays are stable, so useLatestPublications now
memoizes its merged list on the two query results rather than rebuilding
it each render.
