+++ src/containers/globe/live-updates/mobile/carousel.tsx
